Refactor: split DropdownMenu into per-component files
